+++ org.cchmc.pho.core/DataModels/WebChartData.cs
@@ -25,7 +25,7 @@ namespace org.cchmc.pho.core.DataModels
     {
         public string Type { get; set; }
         public string[] XAxisLabels { get; set; }
-        public int[] Values { get; set; }
+        public decimal[] Values { get; set; }
         public string Legend { get; set; }
         public string BackgroundColor { get; set; }
         public string BackgroundHoverColor { get; set; }
